Usersignup dump/restore
