+++ src/Distribution/Server/Features/UserSignup.hs
@@ -10,11 +10,11 @@ module Distribution.Server.Features.UserSignup (
   ) where
 
 import qualified Distribution.Server.Features.UserSignup.Acid as Acid
+import Distribution.Server.Features.UserSignup.Backup
 
 import Distribution.Server.Framework
 import Distribution.Server.Framework.Templating
 import Distribution.Server.Framework.BackupDump
-import Distribution.Server.Framework.BackupRestore
 
 import Distribution.Server.Features.Upload
 import Distribution.Server.Features.Users
@@ -35,7 +35,6 @@ import qualified Data.ByteString.Lazy as BSL
 
 import Distribution.Text (display)
 import Data.Time
-import Text.CSV (CSV, Record)
 import Network.Mail.Mime
 import Network.URI (URI(..), URIAuth(..))
 import Graphics.Captcha
@@ -105,92 +104,6 @@ signupResetStateComponent stateDir = do
     , restoreState = signupResetBackup
     , resetState   = signupResetStateComponent
     }
-
-----------------------------
--- Data backup and restore
---
-
-signupResetBackup :: RestoreBackup Acid.SignupResetTable
-signupResetBackup = go []
-  where
-   go :: [(Nonce, Acid.SignupResetInfo)] -> RestoreBackup Acid.SignupResetTable
-   go st =
-     RestoreBackup {
-       restoreEntry = \entry -> case entry of
-         BackupByteString ["signups.csv"] bs -> do
-           csv <- importCSV "signups.csv" bs
-           signups <- importSignupInfo csv
-           return (go (signups ++ st))
-
-         BackupByteString ["resets.csv"] bs -> do
-           csv <- importCSV "resets.csv" bs
-           resets <- importResetInfo csv
-           return (go (resets ++ st))
-
-         _ -> return (go st)
-
-     , restoreFinalize =
-        return (Acid.SignupResetTable (Map.fromList st))
-     }
-
-importSignupInfo :: CSV -> Restore [(Nonce, Acid.SignupResetInfo)]
-importSignupInfo = mapM fromRecord . drop 2
-  where
-    fromRecord :: Record -> Restore (Nonce, Acid.SignupResetInfo)
-    fromRecord [nonceStr, usernameStr, realnameStr, emailStr, timestampStr] = do
-        timestamp <- parseUTCTime "timestamp" timestampStr
-        nonce <- parseNonceM nonceStr
-        let signupinfo = Acid.SignupInfo {
-              signupUserName     = T.pack usernameStr,
-              signupRealName     = T.pack realnameStr,
-              signupContactEmail = T.pack emailStr,
-              nonceTimestamp     = timestamp
-            }
-        return (nonce, signupinfo)
-    fromRecord x = fail $ "Error processing signup info record: " ++ show x
-
-signupInfoToCSV :: BackupType -> Acid.SignupResetTable -> CSV
-signupInfoToCSV backuptype (Acid.SignupResetTable tbl)
-    = ["0.1"]
-    : [ "token", "username", "realname", "email", "timestamp" ]
-    : [ [ if backuptype == FullBackup
-          then renderNonce nonce
-          else ""
-        , T.unpack signupUserName
-        , T.unpack signupRealName
-        , if backuptype == FullBackup
-          then T.unpack signupContactEmail
-          else "hidden-email@nowhere.org"
-        , formatUTCTime nonceTimestamp
-        ]
-      | (nonce, Acid.SignupInfo{..}) <- Map.toList tbl ]
-
-importResetInfo :: CSV -> Restore [(Nonce, Acid.SignupResetInfo)]
-importResetInfo = mapM fromRecord . drop 2
-  where
-    fromRecord :: Record -> Restore (Nonce, Acid.SignupResetInfo)
-    fromRecord [nonceStr, useridStr, timestampStr] = do
-        userid <- parseText "userid" useridStr
-        timestamp <- parseUTCTime "timestamp" timestampStr
-        nonce <- parseNonceM nonceStr
-        let signupinfo = Acid.ResetInfo {
-              resetUserId    = userid,
-              nonceTimestamp = timestamp
-            }
-        return (nonce, signupinfo)
-    fromRecord x = fail $ "Error processing signup info record: " ++ show x
-
-resetInfoToCSV :: BackupType -> Acid.SignupResetTable -> CSV
-resetInfoToCSV backuptype (Acid.SignupResetTable tbl)
-    = ["0.1"]
-    : [ "token", "userid", "timestamp" ]
-    : [ [ if backuptype == FullBackup
-          then renderNonce nonce
-          else ""
-        , display resetUserId
-        , formatUTCTime nonceTimestamp
-        ]
-      | (nonce, Acid.ResetInfo{..}) <- Map.toList tbl ]
 
 
 ----------------------------------------
